fix: tests: allow SIGURG in test runs
Golang runtime 1.14+ will internally raise SIGURG for their own
purposes; when we see this, we end up passing it down into the child
process.  As there's no way to tell if that's from the runtime or from
the system, we'll blindly pass it in and hope for the best.  In that
case, though, we should expect them to show up in the test; otherwise
we'll get flaky tests as they panic when the mock processes receive an
unexpected signal.