Don't exit container-run until SIGTERM is received, or the main process fails
To avoid terminating early and killing a still running drain script it is
necessary to keep the container-run process running even if the main
process has already exited.

Once container-run receives the SIGTERM signal it will pass it on to the
child processes and then wait until all direct children have exited before
exiting itself. This should allow the processes to terminate cleanly as
long as the grace period has not yet expired.

container-run will also terminate immediately if the main process terminates
with an error. This allows the container to fail and be restarted by k8s.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>